+++ R/pivot-wide.R
@@ -22,6 +22,10 @@
 #'   Defaults to all columns in `data` except for the columns specified in
 #'   `names_from` and `values_from`. Typically used when you have additional
 #'   variables that is directly related.
+#'
+#'   This takes a tidyselect specification, e.g. use `a:c` to select all
+#'   columns from `a` to `c`, `starts_with("prefix")` to select all columns
+#'   starting with "prefix", or `everything()` to select all columns.
 #' @param names_from,values_from A pair of arguments describing which column
 #'   (or columns) to get the name of the output column (`name_from`), and
 #'   which column (or columns) to get the cell values from (`values_from`).
@@ -34,12 +38,20 @@
 #' @param names_prefix String added to the start of every variable name. This is
 #'   particularly useful if `names_from` is a numeric vector and you want to
 #'   create syntactic variable names.
-#' @param values_fill Optionally, a named list specifying what each `value`
+#' @param names_glue Instead of `names_sep` and `names_prefix`, you can supply
+#'   a glue specification that uses the `names_from` columns (and special
+#'   `.value`) to create custom column names.
+#' @param values_fill Optionally, a value that specifies what each `value`
 #'   should be filled in with when missing.
-#' @param values_fn Optionally, a named list providing a function that will be
-#'   applied to the `value` in each cell in the output. You will typically
-#'   use this when the combination of `id_cols` and `value` column does not
-#'   uniquely identify an observation.
+#'
+#'   This can be a named list if you want to apply different aggregations
+#'   to different value columns.
+#' @param values_fn Optionally, a function applied to the `value` in each cell
+#'   in the output. You will typically use this when the combination of
+#'   `id_cols` and `value` column does not uniquely identify an observation.
+#'
+#'   This can be a named list if you want to apply different aggregations
+#'   to different value columns.
 #' @export
 #' @examples
 #' # See vignette("pivot") for examples and explanation
@@ -49,16 +61,27 @@
 #'   pivot_wider(names_from = station, values_from = seen)
 #' # Fill in missing values
 #' fish_encounters %>%
-#'   pivot_wider(
-#'     names_from = station,
-#'     values_from = seen,
-#'     values_fill = list(seen = 0)
-#'   )
+#'   pivot_wider(names_from = station, values_from = seen, values_fill = 0)
 #'
 #' # Generate column names from multiple variables
 #' us_rent_income
 #' us_rent_income %>%
 #'   pivot_wider(names_from = variable, values_from = c(estimate, moe))
+#'
+#' # When there are multiple `names_from` or `values_from`, you can use
+#' # use `names_sep` or `names_glue` to control the output variable names
+#' us_rent_income %>%
+#'   pivot_wider(
+#'     names_from = variable,
+#'     names_sep = ".",
+#'     values_from = c(estimate, moe)
+#'   )
+#' us_rent_income %>%
+#'   pivot_wider(
+#'     names_from = variable,
+#'     names_glue = "{variable}_{.value}",
+#'     values_from = c(estimate, moe)
+#'   )
 #'
 #' # Can perform aggregation with values_fn
 #' warpbreaks <- as_tibble(warpbreaks[c("wool", "tension", "breaks")])
@@ -67,13 +90,14 @@
 #'   pivot_wider(
 #'     names_from = wool,
 #'     values_from = breaks,
-#'     values_fn = list(breaks = mean)
+#'     values_fn = mean
 #'   )
 pivot_wider <- function(data,
                         id_cols = NULL,
                         names_from = name,
                         names_prefix = "",
                         names_sep = "_",
+                        names_glue = NULL,
                         names_repair = "check_unique",
                         values_from = value,
                         values_fill = NULL,
@@ -85,7 +109,8 @@ pivot_wider <- function(data,
     names_from = !!names_from,
     values_from = !!values_from,
     names_prefix = names_prefix,
-    names_sep = names_sep
+    names_sep = names_sep,
+    names_glue = names_glue
   )
 
   id_cols <- enquo(id_cols)
@@ -104,6 +129,43 @@ pivot_wider <- function(data,
 #' @keywords internal
 #' @export
 #' @inheritParams pivot_wider
+#' @param spec A specification data frame. This is useful for more complex
+#'  pivots because it gives you greater control on how metadata stored in the
+#'  columns become column names in the result.
+#'
+#'   Must be a data frame containing character `.name` and `.value` columns.
+#'   Additional columns in `spec` should be named to match columns in the
+#'   long format of the dataset and contain values corresponding to columns
+#'   pivoted from the wide format.
+#'   The special `.seq` variable is used to disambiguate rows internally;
+#'   it is automatically removed after pivotting.
+#'
+#' @examples
+#' # See vignette("pivot") for examples and explanation
+#'
+#' us_rent_income
+#' spec1 <- us_rent_income %>%
+#'   build_wider_spec(names_from = variable, values_from = c(estimate, moe))
+#' spec1
+#'
+#' us_rent_income %>%
+#'   pivot_wider_spec(spec1)
+#'
+#' # Is equivalent to
+#' us_rent_income %>%
+#'   pivot_wider(names_from = variable, values_from = c(estimate, moe))
+#'
+#' # `pivot-wider_spec()` provides more control over column names and output format
+#' # instead of creating columns with estimate_ and moe_ prefixes,
+#' # keep original variable name for estimates and attach _moe as suffix
+#' spec2 <- tibble(
+#'   .name = c("income", "rent", "income_moe", "rent_moe"),
+#'   .value = c("estimate", "estimate", "moe", "moe"),
+#'   variable = c("income", "rent", "income", "rent")
+#' )
+#'
+#' us_rent_income %>%
+#'   pivot_wider_spec(spec2)
 pivot_wider_spec <- function(data,
                                   spec,
                                   names_repair = "check_unique",
@@ -111,6 +173,14 @@ pivot_wider_spec <- function(data,
                                   values_fill = NULL,
                                   values_fn = NULL) {
   spec <- check_spec(spec)
+
+  if (is.function(values_fn)) {
+    values_fn <- rep_named(unique(spec$.value), list(values_fn))
+  }
+
+  if (is_scalar(values_fill)) {
+    values_fill <- rep_named(unique(spec$.value), list(values_fill))
+  }
 
   values <- vec_unique(spec$.value)
   spec_cols <- c(names(spec)[-(1:2)], values)
@@ -127,10 +197,12 @@ pivot_wider_spec <- function(data,
   df_rows <- data[key_vars]
   if (ncol(df_rows) == 0) {
     rows <- tibble(.rows = 1)
+    nrow <- 1L
     row_id <- rep(1L, nrow(df_rows))
   } else {
-    rows <- vec_unique(df_rows)
-    row_id <- vec_match(df_rows, rows)
+    row_id <- vec_group_id(df_rows)
+    nrow <- attr(row_id, "n")
+    rows <- vec_slice(df_rows, vec_unique_loc(row_id))
   }
 
   value_specs <- unname(split(spec, spec$.value))
@@ -142,7 +214,7 @@ pivot_wider_spec <- function(data,
     val <- data[[value]]
 
     cols <- data[names(spec_i)[-(1:2)]]
-    col_id <- vec_match(cols, spec_i[-(1:2)])
+    col_id <- vec_match(as_tibble(cols), spec_i[-(1:2)])
     val_id <- data.frame(row = row_id, col = col_id)
 
     dedup <- vals_dedup(
@@ -154,7 +226,6 @@ pivot_wider_spec <- function(data,
     val_id <- dedup$key
     val <- dedup$val
 
-    nrow <- nrow(rows)
     ncol <- nrow(spec_i)
 
     fill <- values_fill[[value]]
@@ -170,7 +241,7 @@ pivot_wider_spec <- function(data,
     value_out[[i]] <- wrap_vec(out, spec_i$.name)
   }
 
-  out <- wrap_error_names(vec_cbind(rows, !!!value_out, .name_repair = names_repair))
+  out <- wrap_error_names(vec_cbind(as_tibble(rows), !!!value_out, .name_repair = names_repair))
 
   # recreate desired column order
   # https://github.com/r-lib/vctrs/issues/227
@@ -189,7 +260,8 @@ build_wider_spec <- function(data,
                              names_from = name,
                              values_from = value,
                              names_prefix = "",
-                             names_sep = "_") {
+                             names_sep = "_",
+                             names_glue = NULL) {
   names_from <- tidyselect::vars_select(tbl_vars(data), !!enquo(names_from))
   values_from <- tidyselect::vars_select(tbl_vars(data), !!enquo(values_from))
 
@@ -210,7 +282,12 @@ build_wider_spec <- function(data,
     row_ids <- vec_repeat(row_ids, times = vec_size(values_from))
   }
 
-  vec_cbind(out, row_ids)
+  out <- vec_cbind(out, as_tibble(row_ids), .name_repair = "minimal")
+  if (!is.null(names_glue)) {
+    out$.name <- as.character(glue::glue_data(out, names_glue))
+  }
+
+  out
 }
 
 # quiet R CMD check
@@ -227,17 +304,17 @@ vals_dedup <- function(key, val, value, summarize = NULL) {
     }
 
     warn(glue::glue(
-      "Values in `{value}` are not uniquely identified; output will contain list-cols.\n",
-      "* Use `values_fn = list({value} = list)` to suppress this warning.\n",
-      "* Use `values_fn = list({value} = length)` to identify where the duplicates arise\n",
-      "* Use `values_fn = list({value} = summary_fun)` to summarise duplicates"
+      "Values are not uniquely identified; output will contain list-cols.\n",
+      "* Use `values_fn = list` to suppress this warning.\n",
+      "* Use `values_fn = length` to identify where the duplicates arise\n",
+      "* Use `values_fn = {{summary_fun}}` to summarise duplicates"
     ))
   }
 
   out <- vec_split(val, key)
   if (!is.null(summarize) && !identical(summarize, list)) {
     summarize <- as_function(summarize)
-    # This is only correct if `values_collapse` always returns a single value
+    # This is only correct if `values_fn` always returns a single value
     # Needs https://github.com/r-lib/vctrs/issues/183
     out$val <- vec_c(!!!map(out$val, summarize))
   }
@@ -248,11 +325,23 @@ vals_dedup <- function(key, val, value, summarize = NULL) {
 # Wrap a "rectangular" vector into a data frame
 wrap_vec <- function(vec, names) {
   ncol <- length(names)
-  nrow <- length(vec) / ncol
+  nrow <- vec_size(vec) / ncol
   out <- set_names(vec_init(list(), ncol), names)
   for (i in 1:ncol) {
     out[[i]] <- vec_slice(vec, ((i - 1) * nrow + 1):(i * nrow))
   }
 
   as_tibble(out)
+}
+
+is_scalar <- function(x) {
+  if (is.null(x)) {
+    return(FALSE)
+  }
+
+  if (is.list(x)) {
+    (length(x) == 1) && !have_name(x)
+  } else {
+    TRUE
+  }
 }