+++ tests/testthat/test-expand.R
@@ -13,7 +13,7 @@ test_that("multiple variables in one arg doesn't expand", {
 })
 
 test_that("nesting doesn't expand values", {
-  df <- data.frame(x = 1:2, y = 1:2)
+  df <- tibble(x = 1:2, y = 1:2)
   expect_equal(expand(df, nesting(x, y)), df)
 })
 
@@ -27,7 +27,7 @@ test_that("unnamed data frames are flattened", {
 })
 
 test_that("named data frames are not flattened", {
-  df <- data.frame(x = 1:2, y = 1:2)
+  df <- tibble(x = 1:2, y = 1:2)
   out <- expand(df, x = nesting(x, y))
   expect_equal(out$x, df)
 
@@ -125,4 +125,39 @@ test_that("crossing handles list columns", {
   expect_s3_class(out, "tbl_df")
   expect_equal(out$x, rep(x, each = 2))
   expect_equal(out$y, rep(y, 2))
+})
+
+test_that("expand_grid can control name_repair", {
+  x <- 1:2
+
+  if (packageVersion("tibble") > "2.99") {
+    expect_error(expand_grid(x, x), class = "rlang_error")
+  } else {
+    expect_error(expand_grid(x, x), "must not be duplicated")
+  }
+
+  expect_message(out <- expand_grid(x, x, .name_repair = "unique"), "New names:")
+  expect_named(out, c("x...1", "x...2"))
+
+  out <- expand_grid(x, x, .name_repair = "minimal")
+  expect_named(out, c("x", "x"))
+})
+
+test_that("crossing/nesting/expand respect .name_repair", {
+
+  x <- 1:2
+  expect_named(crossing(x, x, .name_repair = "unique"), c("x...1", "x...2"))
+
+  expect_named(nesting(x, x, .name_repair = "unique"), c("x...1", "x...2"))
+
+  df <- tibble(x)
+  expect_named(df %>% expand(x, x, .name_repair = "unique"), c("x...1", "x...2"))
+})
+
+
+# dots_cols supports lazy evaluation --------------------------------------
+
+test_that("dots_cols evaluates each expression in turn", {
+  out <- dots_cols(x = seq(-2, 2), y = x)
+  expect_equal(out$x, out$y)
 })