+++ tests/testthat/test-rectangle.R
@@ -10,6 +10,8 @@ test_that("hoist extracts named elements", {
 })
 
 test_that("can require specific type with ptype", {
+  skip("restricted vec_cast")
+
   df <- tibble(x = list(
     list(a = 1),
     list(a = "a")
@@ -116,13 +118,23 @@ test_that("simplifies length-1 lists", {
   df <- tibble(
     x = 1:2,
     y = list(
-      list(a = 1, b = 2),
+      list(a = 1, b = 2, c = c(1, 2)),
       list(a = 3)
     )
   )
   out <- df %>% unnest_wider(y)
   expect_equal(out$a, c(1, 3))
   expect_equal(out$b, c(2, NA))
+  expect_equal(out$c, list(c(1, 2), NULL))
+
+  # Works when casting too
+  skip("restricted vec_cast")
+  out <- df %>% unnest_wider(y,
+    ptype = list(a = double(), b = double(), c = list())
+  )
+  expect_equal(out$a, c(1, 3))
+  expect_equal(out$b, c(2, NA))
+  expect_equal(out$c, list(c(1, 2), NULL))
 })
 
 test_that("can handle data frames consistently with vectors" , {
@@ -153,6 +165,19 @@ test_that("list_of columns can be unnested", {
 
   df <- tibble(x = 1:2, y = list_of(c(a = 1L), c(b = 1:2)))
   expect_named(unnest_wider(df, y), c("x", "a", "b1", "b2"))
+})
+
+test_that("names_sep creates unique names", {
+  df <- tibble(
+    x = list("a", c("a", "b", "c")),
+    y = list(c(a = 1), c(b = 2, a = 1))
+  )
+  expect_warning(out <- unnest_wider(df, x, names_sep = "_"), NA)
+  expect_named(out, c("x_1", "x_2", "x_3", "y"))
+
+  expect_warning(out <- unnest_wider(df, y, names_sep = "_"), NA)
+  expect_named(out, c("x", "y_a", "y_b"))
+  expect_equal(out$y_a, c(1, 1))
 })
 
 # unnest_longer -----------------------------------------------------------
@@ -239,3 +264,4 @@ test_that("mix of named and unnamed becomes longer", {
   expect_message(out <- df %>% unnest_auto(y), "unnest_longer")
   expect_named(out, c("x", "y"))
 })
+